Fix circular dependency in Menu
https://github.com/ant-design/ant-design/issues/17803
