+++ components/menu/SubMenu.tsx
@@ -3,7 +3,7 @@ import * as PropTypes from 'prop-types';
 import { SubMenu as RcSubMenu } from 'rc-menu';
 import classNames from 'classnames';
 
-import { MenuContext, MenuContextProps } from './index';
+import MenuContext, { MenuContextProps } from './MenuContext';
 
 interface TitleEventEntity {
   key: string;
