+++ components/menu/index.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
 import RcMenu, { Divider, ItemGroup } from 'rc-menu';
-import createContext from '@ant-design/create-react-context';
 import classNames from 'classnames';
 import omit from 'omit.js';
 import SubMenu from './SubMenu';
@@ -11,6 +10,7 @@ import warning from '../_util/warning';
 import { polyfill } from 'react-lifecycles-compat';
 import { SiderContext, SiderContextProps } from '../layout/Sider';
 import raf from '../_util/raf';
+import MenuContext, { MenuTheme } from './MenuContext';
 
 export interface SelectParam {
   key: string;
@@ -28,8 +28,6 @@ export interface ClickParam {
 }
 
 export type MenuMode = 'vertical' | 'vertical-left' | 'vertical-right' | 'horizontal' | 'inline';
-
-export type MenuTheme = 'light' | 'dark';
 
 export interface MenuProps {
   id?: string;
@@ -73,15 +71,6 @@ export interface MenuState {
   prevProps: InternalMenuProps;
   mounted: boolean;
 }
-
-export interface MenuContextProps {
-  inlineCollapsed: boolean;
-  antdMenuTheme?: MenuTheme;
-}
-
-export const MenuContext = createContext<MenuContextProps>({
-  inlineCollapsed: false,
-});
 
 class InternalMenu extends React.Component<InternalMenuProps, MenuState> {
   static defaultProps: Partial<MenuProps> = {
